First sketch for usage of native_filter_configuration
We make use of the mapping logic of default_filters (deprecated?) to native_filter_configuration for dashboards during a cache warmup.
The transformed filters are then passed as extra_form_data downstream where they are processed correctly.

This fixes an issue where the cache warmup will not use the filters a user sees when opening a dashboard.